requirements.txt
# SPDX-License-Identifier: MIT
# Copyright (c) 2025 Symbolic MCP Contributors

fastmcp>=2.0.0
crosshair-tool>=0.0.70
z3-solver>=4.12.0
icontract>=2.6.0
typing-extensions>=4.0.0
pytest>=7.0.0
bandit>=1.9.0
RestrictedPython>=8.1
pydantic>=2.12.0
psutil>=6.0.0
